+++ Classes/DataProviders/CornerstoneOverviewDataProvider.php
@@ -1,4 +1,5 @@
 <?php
+
 namespace YoastSeoForTypo3\YoastSeo\DataProviders;
 
 /*
@@ -14,6 +15,7 @@ namespace YoastSeoForTypo3\YoastSeo\DataProviders;
  * The TYPO3 project - inspiring people to share!
  */
 
+use Doctrine\DBAL\Driver\ResultStatement;
 use TYPO3\CMS\Core\Database\ConnectionPool;
 use TYPO3\CMS\Core\Utility\GeneralUtility;
 
@@ -22,34 +24,34 @@ use TYPO3\CMS\Core\Utility\GeneralUtility;
  */
 class CornerstoneOverviewDataProvider extends AbstractOverviewDataProvider
 {
-
     /**
      * @param bool $returnOnlyCount
      * @return int|array
      */
     public function getData($returnOnlyCount = false)
     {
-        $language = (int)$this->callerParams['language'];
-        $constraints = [];
-        $table = 'pages';
+        return $this->getRestrictedPagesResults($returnOnlyCount);
+    }
 
-        $qb = GeneralUtility::makeInstance(ConnectionPool::class)->getQueryBuilderForTable($table);
+    protected function getResults(array $pageIds = []): ?ResultStatement
+    {
+        $queryBuilder = GeneralUtility::makeInstance(ConnectionPool::class)->getQueryBuilderForTable(self::PAGES_TABLE);
 
-        if ($language > 0) {
-            $constraints[] = $qb->expr()->eq('sys_language_uid', $language);
+        $constraints = [
+            $queryBuilder->expr()->eq('sys_language_uid', (int)$this->callerParams['language']),
+            $queryBuilder->expr()->eq('tx_yoastseo_cornerstone', 1)
+        ];
+
+        if (count($pageIds) > 0) {
+            $constraints[] = $queryBuilder->expr()->in(
+                (int)$this->callerParams['language'] > 0 ? $GLOBALS['TCA']['pages']['ctrl']['transOrigPointerField'] : 'uid',
+                $pageIds
+            );
         }
 
-        $constraints[] = $qb->expr()->eq('tx_yoastseo_cornerstone', 1);
-
-        $query = $qb->select('*')
-            ->from($table)
+        return $queryBuilder->select('*')
+            ->from(self::PAGES_TABLE)
             ->where(...$constraints)
             ->execute();
-
-        if ($returnOnlyCount === false) {
-            return $query->fetchAll();
-        }
-
-        return $query->rowCount();
     }
 }