+++ Classes/DataProviders/PagesWithoutDescriptionOverviewDataProvider.php
@@ -1,5 +1,7 @@
 <?php
+
 declare(strict_types=1);
+
 namespace YoastSeoForTypo3\YoastSeo\DataProviders;
 
 /*
@@ -15,6 +17,7 @@ namespace YoastSeoForTypo3\YoastSeo\DataProviders;
  * The TYPO3 project - inspiring people to share!
  */
 
+use Doctrine\DBAL\Driver\ResultStatement;
 use TYPO3\CMS\Core\Database\ConnectionPool;
 use TYPO3\CMS\Core\Utility\GeneralUtility;
 use YoastSeoForTypo3\YoastSeo\Utility\YoastUtility;
@@ -24,6 +27,7 @@ use YoastSeoForTypo3\YoastSeo\Utility\YoastUtility;
  */
 class PagesWithoutDescriptionOverviewDataProvider extends AbstractOverviewDataProvider
 {
+    protected const PAGES_TABLE = 'pages';
 
     /**
      * @param bool $returnOnlyCount
@@ -31,30 +35,33 @@ class PagesWithoutDescriptionOverviewDataProvider extends AbstractOverviewDataPr
      */
     public function getData($returnOnlyCount = false)
     {
-        $doktypes = implode(',', YoastUtility::getAllowedDoktypes()) ?: '-1';
-        $table = 'pages';
+        return $this->getRestrictedPagesResults($returnOnlyCount);
+    }
 
-        $qb = GeneralUtility::makeInstance(ConnectionPool::class)->getQueryBuilderForTable($table);
+    protected function getResults(array $pageIds = []): ?ResultStatement
+    {
+        $queryBuilder = GeneralUtility::makeInstance(ConnectionPool::class)->getQueryBuilderForTable(self::PAGES_TABLE);
 
         $constraints = [
-            $qb->expr()->orX(
-                $qb->expr()->eq('description', $qb->createNamedParameter('')),
-                $qb->expr()->isNull('description')
+            $queryBuilder->expr()->orX(
+                $queryBuilder->expr()->eq('description', $queryBuilder->createNamedParameter('')),
+                $queryBuilder->expr()->isNull('description')
             ),
-            $qb->expr()->in('doktype', $doktypes),
-            $qb->expr()->eq('tx_yoastseo_hide_snippet_preview', 0),
-            $qb->expr()->eq('sys_language_uid', (int)$this->callerParams['language'])
+            $queryBuilder->expr()->in('doktype', YoastUtility::getAllowedDoktypes()),
+            $queryBuilder->expr()->eq('tx_yoastseo_hide_snippet_preview', 0),
+            $queryBuilder->expr()->eq('sys_language_uid', (int)$this->callerParams['language'])
         ];
 
-        $query = $qb->select('*')
-            ->from($table)
-            ->where(...$constraints)
-            ->execute();
-
-        if ($returnOnlyCount === false) {
-            return $query->fetchAll();
+        if (count($pageIds) > 0) {
+            $constraints[] = $queryBuilder->expr()->in(
+                (int)$this->callerParams['language'] > 0 ? $GLOBALS['TCA']['pages']['ctrl']['transOrigPointerField'] : 'uid',
+                $pageIds
+            );
         }
 
-        return $query->rowCount();
+        return $queryBuilder->select(...self::PAGES_FIELDS)
+            ->from(self::PAGES_TABLE)
+            ->where(...$constraints)
+            ->execute();
     }
 }