+++ CHANGELOG.md
@@ -8,6 +8,13 @@ We will follow [Semantic Versioning](http://semver.org/).
 ## Yoast SEO Premium for TYPO3
 Besides the free version of our plugin, we also have a premium version. The free version enables you to do all necessary optimizations. With the premium version, we make it even easier to do! More information can be found on https://www.maxserv.com/yoast.
 
+## UNRELEASED
+### Breaking
+- Dropped support for PHP <7.4
+
+### Added
+- New feature to automatically activate Yoast SEO functionality on custom records
+
 ## 8.2.0 June 23, 2022
 ### Added
 - Support for new TYPO3 composer mode